Initial import of fauxlizer parser
This parses the data generated by Fauxlizer 5000, created
at the Fantasy Analytics lab. The parser handles buggy output
that may be generated from time to time.

* rows may be truncated
* the header may be missing or malformed
* the values in a row may be in the wrong order
* there might not be any data

--
extract_data(file_name)
this will ensure that the required headers
"experiment_name", "sample_id", "fauxness", "category_guess", are
in the csv file. It will then check that:

 * "experiment_name" column is not empty.
 * "sample_id" column has a positive integer ID
 * "fauxness" column is a floating point value on the range [0.0, 1.0]
   inclusive
 * category_guess column is a string enum chosen from the set {"real",
  "fake", "ambiguous"}

this function returns a tuple of (code, results) where
 * "code": is the return code of the function. It can be SUCCESS, INVALID_HEADERS, EMPTY_EXPERIMENT_NAME, SAMPLE_ID_NOT_INT, SAMPLE_ID_NEGATIVE, FAUXNESS_NOT_FLOAT, FAUXNESS_OUT_OF_RANGE, INVALID_CATEGORY_GUESS, NO_DATA.
 * "results": if code is SUCCESS, this is the data from the file. otherwise, it returns
the invalid row from the file.

--
generate_summary(code: str, payload: dict)
this function generates a summary
of the data. "code" is the return code from extract_data().
If code is "SUCCESS", this function will return a summary of the data.
Otherwise, it will return the error code and the payload, which will be the
faulty data that extract_data() detected.

--
fetch_row(payload, row_num, format)
this will return one row in the file
with the format of "format". By default, the returned data will be the
python dict of the row. Current valid formats are "JSON" and "CSV".