Turn program into script and add argument parsing
The code is now usable from the command line to info from data file.

The script takes a positional argument of a file_name. Optionally, the
user can pass in
- a format argument -f for CSV and JSON. If format is not
specificed, it will return the output formatted as a python dict.
- a linenum argument -l. This will return the data at line number in the
  file. linenum is zero indexed.

The summary is always returned and formatted in JSON.